fix reset theme logic on property control
