+++ app/client/src/widgets/JSONFormWidget/widget/propertyConfig/helper.ts
@@ -10,18 +10,12 @@ import {
   HookResponse,
   FieldThemeStylesheet,
   ROOT_SCHEMA_KEY,
-  getBindingTemplate,
 } from "../../constants";
 import { getGrandParentPropertyPath, getParentPropertyPath } from "../helper";
 import { JSONFormWidgetProps } from "..";
 import { getFieldStylesheet } from "widgets/JSONFormWidget/helper";
 import { AppTheme } from "entities/AppTheming";
 import { processSchemaItemAutocomplete } from "components/propertyControls/JSONFormComputeControl";
-import {
-  combineDynamicBindings,
-  getDynamicBindings,
-  isDynamicValue,
-} from "utils/DynamicBindingUtils";
 
 export type HiddenFnParams = [JSONFormWidgetProps, string];
 
@@ -111,28 +105,15 @@ export const getStylesheetValue = (
   propertyPath: string,
   widgetStylesheet?: AppTheme["stylesheet"][string],
 ) => {
-  return getSchemaItem(props, propertyPath).compute(
-    (schemaItem, propertyName) => {
-      const fieldStylesheet = getFieldStylesheet(
-        schemaItem.fieldType,
-        widgetStylesheet?.childStylesheet as FieldThemeStylesheet,
-      );
+  return getSchemaItem(props, propertyPath).compute((schemaItem) => {
+    const fieldStylesheet = getFieldStylesheet(
+      props.widgetName,
+      schemaItem.fieldType,
+      widgetStylesheet?.childStylesheet as FieldThemeStylesheet,
+    );
 
-      if (isDynamicValue(fieldStylesheet[propertyName])) {
-        const { jsSnippets, stringSegments } = getDynamicBindings(
-          fieldStylesheet[propertyName],
-        );
-        const js = combineDynamicBindings(jsSnippets, stringSegments);
-        const { prefixTemplate, suffixTemplate } = getBindingTemplate(
-          props.widgetName,
-        );
-
-        return `${prefixTemplate}${js}${suffixTemplate}`;
-      }
-
-      return fieldStylesheet || "";
-    },
-  );
+    return fieldStylesheet || "";
+  });
 };
 
 export const getAutocompleteProperties = (props: JSONFormWidgetProps) => {
